chore(TCOMP-681): extract common part from action classes to base class
